% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_values.R
\name{rand_values}
\alias{rand_values}
\title{A Cat Function}
\usage{
rand_values(n = 50)
}
\arguments{
\item{n}{Number of random numbers}
}
\description{
This function allows you to get n random numbers
}
\examples{
rand_values()
}
\keyword{numbers}
\keyword{random}

